% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admin1}
\alias{admin1}
\title{Administrative division 1 data...}
\format{
A data frame with 3956 rows and 4 variables:
\describe{
  \item{code}{concatenated admin code}
  \item{iso}{ISO 3166-1 alpha-2 code}
  \item{lvl.1}{level 1 admin subdivision code}
  \item{name}{name}
  ...
}
}
\source{
\url{http://download.geonames.org/export/dump/admin1CodesASCII.txt}
}
\usage{
admin1
}
\description{
A dataset containing administrative division 1 data.
}
\keyword{datasets}
