% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admin2}
\alias{admin2}
\title{Administrative division 2 data...}
\format{
A data frame with 44737 rows and 5 variables:
\describe{
  \item{code}{concatenated admin code}
  \item{iso}{ISO 3166-1 alpha-2 code}
  \item{lvl.1}{level 1 admin subdivision code}
  \item{lvl.2}{level 2 admin subdivision code}
  \item{name}{name}
  ...
}
}
\source{
\url{http://download.geonames.org/export/dump/admin2Codes.txt}
}
\usage{
admin2
}
\description{
A dataset containing administrative division 2 data.
}
\keyword{datasets}
