% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neogeonames.R
\name{adminify}
\alias{adminify}
\title{Calculate administrative division codes by splitting on a delimiter pattern.}
\usage{
adminify(
  query,
  delim,
  dfgeo = neogeonames::geoname,
  dfcou = neogeonames::country,
  ...
)
}
\arguments{
\item{query}{The place name query.}

\item{delim}{The delimiter pattern.}

\item{dfgeo}{The data frame of \code{\link{geoname}} data.}

\item{dfcou}{The data frame of \code{\link{country}} data.}

\item{...}{The arguments passed to \code{\link{countrify}} and \code{\link{geonamify}}.}
}
\value{
The list with "id" and "ac" atomic vectors consisting of the geonameid and
        administrative class division values or \code{NA} values if missing.
}
\description{
This function infers an admin code and geonameid for each token.
}
\seealso{
\code{\link{geonamify}}
}
