% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neogeonames.R
\name{adminify_delim}
\alias{adminify_delim}
\title{Calculate administrative division codes using a delimiter.}
\usage{
adminify_delim(
  query,
  delim = NA,
  admin = akac,
  n = 1,
  p = list(ignore.case = T)
)
}
\arguments{
\item{query}{The place name query.}

\item{delim}{The delimiter pattern.}

\item{admin}{The admin codes to search.}

\item{n}{The number of allowable fuzzy search results before returning the top result,
otherwise nothing.}

\item{p}{The parameters for \code{\link{agrep}}.}
}
\value{
The list with "id" and "ac" atomic vectors consisting of the geonameid and
        administrative class division values or \code{NA} values if missing.
}
\description{
This function rotates the token vector, returning the result with the most results.
}
\seealso{
\code{\link{adminify_tokens}}
}
