% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neogeonames.R
\name{adminify_regex}
\alias{adminify_regex}
\title{Calculate administrative division codes using a named regular expression.}
\usage{
adminify_regex(query, regex, n = 1, p = list(ignore.case = T))
}
\arguments{
\item{query}{The place name query.}

\item{regex}{The list object with a "pattern" and admin code "name" entry. The pattern entry is
a regular expression with groups that correspond to the names in the name vector.}

\item{n}{The number of allowable fuzzy search results before returning the top result,
otherwise nothing.}

\item{p}{The parameters for \code{\link{agrep}}.}
}
\value{
The list with "id" and "ac" atomic vectors consisting of the geonameid and
        administrative class division values or \code{NA} values if missing.
}
\description{
Only return a result if the all named patterns are found.
}
\seealso{
\code{\link{adminify_regexes}}
}
