% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubricity.R
\name{adminify_regexes}
\alias{adminify_regexes}
\title{Calculate administrative division codes using a list of named regular expression lists.}
\usage{
adminify_regexes(query, regexes, ...)
}
\arguments{
\item{query}{The query.}

\item{regexes}{The list of named regular expression list objects.}

\item{...}{The arguments passed to \code{\link{countrify}} and \code{\link{geonamify}}.}
}
\value{
The list of extracted admin code names or \code{NA} for each name not found.
}
\description{
This function essentially runs \code{adminify_regex} with each regular expression on the value.
}
\seealso{
\code{\link{adminify_regex}}
}
