% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neogeonames.R
\name{adminify_tokens}
\alias{adminify_tokens}
\title{Calculate an administrative division codes for each token according to hierarchy.}
\usage{
adminify_tokens(tokens, ...)
}
\arguments{
\item{tokens}{The place name query tokens.}

\item{...}{The arguments passed to \code{\link{countrify}} and \code{\link{geonamify}}.}
}
\value{
The list with "id" and "ac" atomic vectors consisting of the geonameid and
        administrative class division values or \code{NA} values if missing.
}
\description{
This function infers an admin code and geonameid for each token in order from left-to-right.
}
\seealso{
\code{\link{adminify}}
}
