% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neogeonames.R
\docType{data}
\name{akac}
\alias{akac}
\title{Map admin key to admin column...}
\format{
An object of class \code{character} of length 5.
}
\usage{
akac
}
\description{
Map admin key to admin column...
}
\keyword{datasets}
