% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neogeonames.R
\docType{data}
\name{akfc}
\alias{akfc}
\title{Map admin key to feature codes...}
\format{
An object of class \code{list} of length 5.
}
\usage{
akfc
}
\description{
Map admin key to feature codes...
}
\keyword{datasets}
