% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alternate}
\alias{alternate}
\title{alternate}
\format{
A data frame with 37589 rows and 10 columns
\describe{
  \item{alternateid}{the id of this alternate name}
  \item{geonameid}{geonameId referring to id in table 'geoname'}
  \item{isolanguage}{iso 639 language code 2- or 3-characters; 4-characters 'post' for postal codes and 'iata','icao' and faac for airport codes, fr_1793 for French Revolution names, abbr for abbreviation, link to a website (mostly to wikipedia), wkdt for the wikidataid}
  \item{alternateName}{alternate name or name variant}
  \item{isPreferredName}{if this alternate name is an official/preferred name}
  \item{isShortName}{if this is a short name like 'California' for 'State of California'}
  \item{isColloquial}{if this alternate name is a colloquial or slang term. Example: 'Big Apple' for 'New York'}
  \item{isHistoric}{if this alternate name is historic and was used in the past. Example 'Bombay' for 'Mumbai'}
  \item{from}{from period when the name was used}
  \item{to}{to period when the name was used}
}
}
\source{
\url{http://download.geonames.org/export/dump/alternateNamesV2.zip}
}
\usage{
alternate
}
\description{
This is a data frame that maps alternate names to geoname identifiers.
}
\details{
It is a subset corresponding to the "PCLI", "PCLH", "PCLS", "PCLD", "PCLF", "PCL", and "TERR" feature codes.
}
\keyword{datasets}
