% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubricity.R
\docType{data}
\name{ccode}
\alias{ccode}
\title{Named regex code to admin column code.}
\format{
An object of class \code{character} of length 5.
}
\usage{
ccode
}
\description{
Named regex code to admin column code.
}
\keyword{datasets}
