% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubricity.R
\name{codify.1}
\alias{codify.1}
\title{Calculate the corresponding level 1 admin code for the given query.}
\usage{
codify.1(val, cc2 = NA, n = 1, ...)
}
\arguments{
\item{val}{The query, which corresponds to the level of a state (in the United States for example).}

\item{cc2}{The ISO 3166-1 alpha-2 code to limit the query.}

\item{n}{The number of allowable fuzzy search results before returning the top result, otherwise return nothing if exceeded.}

\item{...}{The arguments for \code{\link{agrep}}.}
}
\value{
The level 1 admin code or \code{NA}.
}
\description{
The function performs a case-insensitive search for an exact match to the country name or ISO codes.
It also performs a fuzzy search using \code{\link{agrep}} as a fall back.
}
