% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubricity.R
\name{countrify}
\alias{countrify}
\title{Calculate the corresponding ISO 3166-1 alpha-2 code for the given query..}
\usage{
countrify(val, n = 1, ...)
}
\arguments{
\item{val}{The country name query.}

\item{n}{The number of allowable fuzzy search results before returning the top result, otherwise return nothing if exceeded.}

\item{...}{The arguments for \code{\link{agrep}}.}
}
\value{
The ISO 3166-1 alpha-2 code or \code{NA}.
}
\description{
The function performs a case-insensitive search for an exact match to the country name or ISO codes.
It also performs a fuzzy search using \code{\link{agrep}} as a fall back.
}
