% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country}
\alias{country}
\title{Country data...}
\format{
A data frame with 252 rows and 3 variables:
\describe{
  \item{iso}{ISO 3166-1 alpha-2 code}
  \item{iso3}{ISO 3166-1 alpha-3 code}
  \item{name}{name}
  ...
}
}
\source{
\url{http://download.geonames.org/export/dump/countryInfo.txt}
}
\usage{
country
}
\description{
A dataset containing country information.
}
\keyword{datasets}
