% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geoname}
\alias{geoname}
\title{The geoname data...}
\format{
A data frame with 400189 rows and 19 columns
\describe{
 \item{geonameid}{geonameid}
 \item{name}{name}
 \item{asciiname}{asciiname}
 \item{alternatenames}{alternatenames}
 \item{latitude}{latitude}
 \item{longitude}{longitude}
 \item{feature_class}{feature_class}
 \item{feature_code}{feature_code}
 \item{country_code}{country_code}
 \item{cc2}{cc2}
 \item{admin1_code}{admin1_code}
 \item{admin2_code}{admin2_code}
 \item{admin3_code}{admin3_code}
 \item{admin4_code}{admin4_code}
 \item{population}{population}
 \item{elevation}{elevation}
 \item{dem}{dem}
 \item{timezone}{timezone}
 \item{modification_date}{modification_date}
  ...
}
}
\source{
\url{http://download.geonames.org/export/dump/allCountries.zip}
}
\usage{
geoname
}
\description{
A dataset containing administrative division 1 data.
}
\keyword{datasets}
