% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubricity.R
\name{geonames_download}
\alias{geonames_download}
\title{Download the raw data from GeoNames Gazetteer.}
\usage{
geonames_download(
  root = getwd(),
  url = "http://download.geonames.org/export/dump"
)
}
\arguments{
\item{root}{The root directory to download the files to.}

\item{url}{The base url to download the files from.}
}
\value{
Returns a character vector of the downloaded paths named by the file name.
}
\description{
Download the countryInfo.txt, admin1CodesASCII.txt, and admin2Codes.txt files.
}
