% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubricity.R
\name{geonames_install}
\alias{geonames_install}
\title{Download and install the GeoNames Gazetteer data.}
\usage{
geonames_install(url = "http://download.geonames.org/export/dump")
}
\arguments{
\item{url}{The base url to download the files from.}
}
\value{
The data frame.
}
\description{
Download and install the GeoNames Gazetteer data.
}
