% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubricity.R
\name{geonames_read_admin_2}
\alias{geonames_read_admin_2}
\title{Read the raw data from the level 2 admin code text file.}
\usage{
geonames_read_admin_2(path)
}
\arguments{
\item{path}{The path to the file.}
}
\value{
The data frame.
}
\description{
Read the raw data from the level 2 admin code text file.
}
