% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubricity.R
\name{geonames_read_country}
\alias{geonames_read_country}
\title{Read the raw data from the country text file.}
\usage{
geonames_read_country(path)
}
\arguments{
\item{path}{The path to the file.}
}
\value{
The data frame.
}
\description{
Read the raw data from the country text file.
}
