% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neogeonames.R
\name{geonamify}
\alias{geonamify}
\title{Calculate the \code{\link{geoname}} rows.}
\usage{
geonamify(
  query,
  dfgeo = neogeonames::geoname,
  where = NULL,
  n = 1,
  p = list(ignore.case = T)
)
}
\arguments{
\item{query}{The place name query.}

\item{dfgeo}{The data frame of \code{\link{geoname}} data.}

\item{where}{The named vector of values analogous to the SQL "WHERE" clause.}

\item{n}{The number of allowable fuzzy search results before returning the top result,
otherwise return nothing if exceeded.}

\item{p}{The parameters for \code{\link{agrep}}.}
}
\value{
The rows or \code{data.frame} with 0 rows.
}
\description{
The function performs a case-insensitive search for matches to the asciiname column.
}
\seealso{
\code{\link{agrep}}
}
