% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neogeonames.R
\name{geonamify_delim}
\alias{geonamify_delim}
\title{Calculate administrative division codes by splitting on a delimiter pattern.}
\usage{
geonamify_delim(query, delim, ...)
}
\arguments{
\item{query}{The query.}

\item{delim}{The delimiter pattern.}

\item{...}{The arguments passed to \code{\link{countrify}} and \code{\link{geonamify}}.}
}
\value{
The list with an "id" and "ac" list consisting of the geonameid and administrative
        division values or \code{NA} if missing.
}
\description{
This function infers an admin code for each token.
}
\seealso{
\code{\link{geonamify}}
}
