% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubricity.R
\name{is.int0}
\alias{is.int0}
\title{Helper function to test for equivalency to 0.}
\usage{
is.int0(val)
}
\arguments{
\item{val}{The value to test.}
}
\value{
The boolean result.
}
\description{
Helper function to test for equivalency to 0.
}
