% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubricity.R
\name{is.prefix}
\alias{is.prefix}
\title{Is y a prefix of x?}
\usage{
is.prefix(x, y)
}
\arguments{
\item{x}{The x value.}

\item{y}{The y value.}
}
\value{
The boolean result.
}
\description{
Is y a prefix of x?
}
