% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{language}
\alias{language}
\title{language}
\format{
A data frame with 7929 rows and 4 columns
\describe{
  \item{iso3}{the ISO 639-3 code}
  \item{iso2}{the ISO 639-2 code}
  \item{iso1}{the ISO 639-1 code}
  \item{name}{the language name}
}
}
\source{
\url{http://download.geonames.org/export/dump/iso-languagecodes.txt}
}
\usage{
language
}
\description{
This is a data frame of ISO 639 language codes, as used for alternate names in the alternate data frame.
}
\keyword{datasets}
