% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubricity.R
\docType{package}
\name{lubricity}
\alias{lubricity}
\alias{lubricity-package}
\title{lubricity: Sniff and Normalize Place Names in R}
\description{
lubricity: Sniff and Normalize Place Names in R

This package provides functions to extract and normalize place names according to administrative division.
}
\author{
\strong{Maintainer}: Daniel Negrón \email{dnegron2@gmu.edu}

}
