% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neogeonames.R
\docType{package}
\name{neogeonames}
\alias{neogeonames}
\alias{neogeonames-package}
\title{neogeonames: Parse, Sniff, and Normalize Place Names in R}
\description{
neogeonames: Parse, Sniff, and Normalize Place Names in R

This package provides a subset of GeoNames Gazetteer data and normalization functions.
}
\author{
\strong{Maintainer}: Daniel Negrón \email{dnegron2@gmu.edu}

}
