% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neogeonames_updated}
\alias{neogeonames_updated}
\title{neogeonames_updated}
\format{
An object of class \code{character} of length 1.
}
\usage{
neogeonames_updated
}
\description{
This is the date when the latest package data download occurred.
}
\keyword{datasets}
