% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubricity.R
\name{process_delim}
\alias{process_delim}
\title{Calculate administrative division codes by splitting on a delimiter pattern.}
\usage{
process_delim(val, pattern, ...)
}
\arguments{
\item{val}{The query value.}

\item{...}{The arguments passed to \code{\link{countrify}}, \code{\link{codify.1}}, and \code{\link{codify.2}}.}

\item{pat}{The delimiter pattern.}
}
\value{
The list of extracted admin code names or \code{NA} for each name not found.
}
\description{
This function infers an admin code for each token.
This function also nullifies results with incorrectly inferred hierarchy.
}
\seealso{
\code{\link{process_regx}}
}
