% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubricity.R
\name{process_regx}
\alias{process_regx}
\title{Calculate administrative division codes using a named regular expression.}
\usage{
process_regx(val, regx, ...)
}
\arguments{
\item{val}{The query value.}

\item{regx}{The list object with a "pattern" and admin code "name" entry.}

\item{...}{The arguments passed to \code{\link{countrify}}, \code{\link{codify.1}}, and \code{\link{codify.2}}.}
}
\value{
The list of extracted admin code names or \code{NA} for each name not found.
}
\description{
The regular expression names each group by the administrative division code.
The function processes each group according to the division hierarchy, starting at the top.
The function uses matches to higher levels as limiters if lower admin code groups are present in the regular expression.
The admin code names include "cn", "ac2", and "ac2" corresponding to "country name", "admin code 1", and "admin code 2".
}
