% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubricity.R
\name{process_regx_list}
\alias{process_regx_list}
\title{Calculate administrative division codes using a list of named regular expression lists.}
\usage{
process_regx_list(val, regx_list, ...)
}
\arguments{
\item{val}{The query value.}

\item{regx_list}{The list of named regular expression list objects.}

\item{...}{The arguments passed to \code{\link{countrify}}, \code{\link{codify.1}}, and \code{\link{codify.2}}.}
}
\value{
The list of extracted admin code names or \code{NA} for each name not found.
}
\description{
This function essentially runs \code{process_regx} on each value.
}
\seealso{
\code{\link{process_regx}}
}
