% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubricity.R
\docType{data}
\name{rcode}
\alias{rcode}
\title{Named regex code to feature admin code.}
\format{
An object of class \code{character} of length 4.
}
\usage{
rcode
}
\description{
Named regex code to feature admin code.
}
\keyword{datasets}
