% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shape}
\alias{shape}
\title{shape}
\format{
A data frame with 191449 rows and 7 columns
\describe{
  \item{geonameid}{the geoname identifier}
  \item{long}{the longitude}
  \item{lat}{the latitude}
  \item{order}{the order}
  \item{hole}{is hole}
  \item{piece}{the piece}
  \item{group}{the group id}
}
}
\source{
\url{http://download.geonames.org/export/dump/shapes_simplified_low.json.zip}
}
\usage{
shape
}
\description{
The simplified country boundaries.
}
\keyword{datasets}
