% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timezone}
\alias{timezone}
\title{timezone}
\format{
A data frame with 425 rows and 5 columns
\describe{
  \item{country}{the country code}
  \item{timezone}{the timezone id}
  \item{gmt}{the GMT offset on 1st of January}
  \item{dst}{the DST offset to gmt on 1st of July (of the current year)}
  \item{rawOffset}{the raw offset without DST}
}
}
\source{
\url{http://download.geonames.org/export/dump/timezone.txt}
}
\usage{
timezone
}
\description{
This is a data frame of time zone data.
}
\keyword{datasets}
