% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neogeonames.R
\name{adminify}
\alias{adminify}
\title{Calculate an administrative division codes for each token according to hierarchy using a delimiter.}
\usage{
adminify(query, delim = " ", admin = akac, n = 1, p = list(ignore.case = T))
}
\arguments{
\item{query}{The place name query.}

\item{delim}{The delimiter pattern.}

\item{admin}{The admin codes to search.}

\item{n}{The number of allowable fuzzy search results before returning the top result,
otherwise nothing.}

\item{p}{The parameters for \code{\link{agrep}}.}
}
\value{
The list with "id" and "ac" atomic vectors consisting of the geonameid and
        administrative class division values or \code{NA} values if missing.
}
\description{
This function infers an admin code and geonameid for each token in order from left-to-right.
}
\seealso{
\code{\link{adminify_tokens}}
}
