% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neogeonames.R
\name{adminify_regexes}
\alias{adminify_regexes}
\title{Calculate administrative division codes using a list of named regular expression.}
\usage{
adminify_regexes(query, regexes, n = 1, p = list(ignore.case = T))
}
\arguments{
\item{query}{The place name query.}

\item{regexes}{The list of named regular expressions.}

\item{n}{The number of allowable fuzzy search results before returning the top result,
otherwise nothing.}

\item{p}{The parameters for \code{\link{agrep}}.}
}
\value{
The list with "id" and "ac" atomic vectors consisting of the geonameid and
        administrative class division values or \code{NA} values if missing.
}
\description{
Only return a result if the all named patterns are found.
}
\seealso{
\code{\link{adminify_regex}}
}
