% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubricity.R
\name{countrify}
\alias{countrify}
\title{Calculate the corresponding row in the country table for the given query.}
\usage{
countrify(query, n = 1, ...)
}
\arguments{
\item{query}{The country name query.}

\item{n}{The number of allowable fuzzy search results before returning the top result, otherwise nothing.}

\item{...}{The parameters for \code{\link{agrep}}.}
}
\value{
The rows or \code{data.frame} with 0 rows.
}
\description{
The function performs a case-insensitive search for matches to ISO codes or country name.
It also performs a fuzzy search using \code{\link{agrep}} as a fall back.
}
\seealso{
\code{\link{agrep}}
}
