% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neogeonames.R
\name{countrify}
\alias{countrify}
\title{Calculate the \code{\link{country}} row.}
\usage{
countrify(
  query,
  dfcou = neogeonames::country,
  n = 1,
  p = list(ignore.case = T)
)
}
\arguments{
\item{query}{The country query.}

\item{dfcou}{The data frame of \code{\link{country}} data.}

\item{n}{The number of allowable fuzzy search results before returning the top result,
otherwise nothing.}

\item{p}{The parameters for \code{\link{agrep}}.}
}
\value{
The rows or \code{data.frame} with 0 rows.
}
\description{
The function performs a case-insensitive search for matches to ISO codes or country name.
It also performs a fuzzy search using \code{\link{agrep}} as a fall back.
}
\seealso{
\code{\link{agrep}}
}
