% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country}
\alias{country}
\title{The country data...}
\format{
A data frame with 252 rows and 19 variables:
\describe{
 \item{iso}{iso}
 \item{iso3}{iso3}
 \item{iso_numeric}{iso_numeric}
 \item{fips}{fips}
 \item{country}{country}
 \item{capital}{capital}
 \item{area}{area}
 \item{population}{population}
 \item{continent}{continent}
 \item{tld}{tld}
 \item{currency_code}{currency_code}
 \item{currency_name}{currency_name}
 \item{phone}{phone}
 \item{postal_code_format}{postal_code_format}
 \item{postal_code_regex}{postal_code_regex}
 \item{languages}{languages}
 \item{geonameid}{geonameid}
 \item{neighbours}{neighbours}
 \item{equivalent_fips_code}{equivalent_fips_code}
  ...
}
}
\source{
\url{http://download.geonames.org/export/dump/countryInfo.txt}
}
\usage{
country
}
\description{
A dataset containing country information.
}
\keyword{datasets}
