% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country}
\alias{country}
\title{country}
\format{
A data frame with 252 rows and 19 variables:
\describe{
  \item{iso}{ISO 3166-1 alpha-2}
  \item{iso3}{ISO 3166-1 alpha-3}
  \item{iso_numeric}{ISO 3166-1 numeric}
  \item{fips}{FIPS codes}
  \item{country}{country}
  \item{capital}{capital}
  \item{area}{area in km^2}
  \item{population}{population}
  \item{continent}{continent}
  \item{tld}{top-level domain}
  \item{currency_code}{ISO 4217}
  \item{currency_name}{currency name}
  \item{phone}{calling code}
  \item{postal_code_format}{postal code format}
  \item{postal_code_regex}{postal code regex}
  \item{languages}{ISO 639 + ISO 3166}
  \item{geonameid}{geonameid}
  \item{neighbours}{neighbouring countries}
  \item{equivalent_fips_code}{equivalent fips code}
}
}
\source{
\url{http://download.geonames.org/export/dump/countryInfo.txt}
}
\usage{
country
}
\description{
This is a data frame of the GeoNames Gazetteer country info data.
}
\details{
CountryCodes:

============

The official ISO country code for the United Kingdom is 'GB'. The code 'UK' is reserved.

A list of dependent countries is available here:
\url{https://spreadsheets.google.com/ccc?key=pJpyPy-J5JSNhe7F_KxwiCA&hl=en}


The countrycode XK temporarily stands for Kosvo:
\url{http://geonames.wordpress.com/2010/03/08/xk-country-code-for-kosovo/}


CS (Serbia and Montenegro) with geonameId = 8505033 no longer exists.
AN (the Netherlands Antilles) with geonameId = 8505032 was dissolved on 10 October 2010.


Currencies :

============

A number of territories are not included in ISO 4217, because their currencies are not per se an independent currency,
but a variant of another currency. These currencies are:

\enumerate{
  \item FO : Faroese krona (1:1 pegged to the Danish krone)
  \item GG : Guernsey pound (1:1 pegged to the pound sterling)
  \item JE : Jersey pound (1:1 pegged to the pound sterling)
  \item IM : Isle of Man pound (1:1 pegged to the pound sterling)
  \item TV : Tuvaluan dollar (1:1 pegged to the Australian dollar).
  \item CK : Cook Islands dollar (1:1 pegged to the New Zealand dollar).
}

The following non-ISO codes are, however, sometimes used: GGP for the Guernsey pound,
JEP for the Jersey pound and IMP for the Isle of Man pound (\url{http://en.wikipedia.org/wiki/ISO_4217})


A list of currency symbols is available here : \url{http://forum.geonames.org/gforum/posts/list/437.page}
another list with fractional units is here: \url{http://forum.geonames.org/gforum/posts/list/1961.page}


Languages :

===========

The column 'languages' lists the languages spoken in a country ordered by the number of speakers. The language code is a 'locale'
where any two-letter primary-tag is an ISO-639 language abbreviation and any two-letter initial subtag is an ISO-3166 country code.

Example : es-AR is the Spanish variant spoken in Argentina.
}
\keyword{datasets}
