% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geoname}
\alias{geoname}
\title{geoname}
\format{
A data frame with 445818 rows and 12 columns
\describe{
  \item{geonameid}{integer id of record in geonames database}
  \item{name}{name of geographical point (utf8) varchar(200)}
  \item{asciiname}{name of geographical point in plain ascii characters, varchar(200)}
  \item{latitude}{latitude in decimal degrees (wgs84)}
  \item{longitude}{longitude in decimal degrees (wgs84)}
  \item{feature_code}{see \url{http://www.geonames.org/export/codes.html}, varchar(10)}
  \item{country_code}{ISO-3166 2-letter country code, 2 characters}
  \item{admin1_code}{fipscode (subject to change to iso code), see exceptions below, see file admin1Codes.txt for display names of this code; varchar(20)}
  \item{admin2_code}{code for the second administrative division, a county in the US, see file admin2Codes.txt; varchar(80) }
  \item{admin3_code}{code for third level administrative division, varchar(20)}
  \item{admin4_code}{code for fourth level administrative division, varchar(20)}
  \item{modification_date}{date of last modification in yyyy-MM-dd format}
}
}
\source{
\url{http://download.geonames.org/export/dump/allCountries.zip}
}
\usage{
geoname
}
\description{
This is a data frame consisting of a subset of the GeoNames Gazetteer all country data.


Feature class:
}
\details{
\tabular{ll}{
  feature_class \tab definition \cr
  A \tab country, state, region, ... \cr
  P \tab city, village, ... \cr
 }

Feature code:

\tabular{llll}{
  feature_class \tab feature_code \tab definition \tab note \cr
  A \tab ADM1 \tab first-order administrative division \tab a primary administrative division of a country, such as a state in the United States \cr
  A \tab ADM1H \tab historical first-order administrative division \tab a former first-order administrative division \cr
  A \tab ADM2 \tab second-order administrative division \tab a subdivision of a first-order administrative division \cr
  A \tab ADM2H \tab historical second-order administrative division \tab a former second-order administrative division \cr
  A \tab ADM3 \tab third-order administrative division \tab a subdivision of a second-order administrative division \cr
  A \tab ADM3H \tab historical third-order administrative division \tab a former third-order administrative division \cr
  A \tab ADM4 \tab fourth-order administrative division \tab a subdivision of a third-order administrative division \cr
  A \tab ADM4H \tab historical fourth-order administrative division \tab a former fourth-order administrative division \cr
  A \tab PCL \tab political entity \tab  \cr
  A \tab PCLD \tab dependent political entity \tab  \cr
  A \tab PCLF \tab freely associated state \tab  \cr
  A \tab PCLH \tab historical political entity \tab a former political entity \cr
  A \tab PCLI \tab independent political entity \tab  \cr
  A \tab PCLS \tab semi-independent political entity \tab  \cr
  A \tab TERR \tab territory \tab  \cr
  P \tab PPLA \tab seat of a first-order administrative division \tab seat of a first-order administrative division (PPLC takes precedence over PPLA) \cr
  P \tab PPLA2 \tab seat of a second-order administrative division \tab  \cr
  P \tab PPLA3 \tab seat of a third-order administrative division \tab  \cr
  P \tab PPLA4 \tab seat of a fourth-order administrative division \tab  \cr
  P \tab PPLC \tab capital of a political entity \tab  \cr
}
}
\keyword{datasets}
