% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geoname}
\alias{geoname}
\title{The geoname data set.}
\format{
A data frame with 445722 rows and 19 columns
\describe{
 \item{geonameid}{integer id of record in geonames database}
 \item{name}{name of geographical point (utf8) varchar(200)}
 \item{asciiname}{name of geographical point in plain ascii characters, varchar(200)}
 \item{alternatenames}{alternatenames, comma separated, ascii names automatically transliterated, convenience attribute from alternatename table, varchar(10000)}
 \item{latitude}{latitude in decimal degrees (wgs84)}
 \item{longitude}{longitude in decimal degrees (wgs84)}
 \item{feature_class}{see <http://www.geonames.org/export/codes.html>, char(1)}
 \item{feature_code}{see <http://www.geonames.org/export/codes.html>, varchar(10)}
 \item{country_code}{ISO-3166 2-letter country code, 2 characters}
 \item{cc2}{alternate country codes, comma separated, ISO-3166 2-letter country code, 200 characters}
 \item{admin1_code}{fipscode (subject to change to iso code), see exceptions below, see file admin1Codes.txt for display names of this code; varchar(20)}
 \item{admin2_code}{code for the second administrative division, a county in the US, see file admin2Codes.txt; varchar(80) }
 \item{admin3_code}{code for third level administrative division, varchar(20)}
 \item{admin4_code}{code for fourth level administrative division, varchar(20)}
 \item{population}{bigint (8 byte int) }
 \item{elevation}{in meters, integer}
 \item{dem}{digital elevation model, srtm3 or gtopo30, average elevation of 3''x3'' (ca 90mx90m) or 30''x30'' (ca 900mx900m) area in meters, integer. srtm processed by cgiar/ciat.}
 \item{timezone}{the iana timezone id (see file <http://download.geonames.org/export/dump/timeZones.txt>) varchar(40)}
 \item{modification_date}{date of last modification in yyyy-MM-dd format}
  ...
}
}
\source{
\url{http://download.geonames.org/export/dump/allCountries.zip}
}
\usage{
geoname
}
\description{
|feature_code|definition|
|-|-|
|A|country, state, region,...|
|P|city, village,...|
}
\details{
|feature_code|feature_code|definition|note|
|-|-|-|-|
|A|ADM1|first-order administrative division|a primary administrative division of a country, such as a state in the United States|
|A|ADM1H|historical first-order administrative division|a former first-order administrative division|
|A|ADM2|second-order administrative division|a subdivision of a first-order administrative division|
|A|ADM2H|historical second-order administrative division|a former second-order administrative division|
|A|ADM3|third-order administrative division|a subdivision of a second-order administrative division|
|A|ADM3H|historical third-order administrative division|a former third-order administrative division|
|A|ADM4|fourth-order administrative division|a subdivision of a third-order administrative division|
|A|ADM4H|historical fourth-order administrative division|a former fourth-order administrative division|
|A|PCL|political entity||
|A|PCLD|dependent political entity||
|A|PCLF|freely associated state||
|A|PCLH|historical political entity|a former political entity|
|A|PCLI|independent political entity||
|A|PCLS|semi-independent political entity||
|A|TERR|territory||
|P|PPLA|seat of a first-order administrative division|seat of a first-order administrative division (PPLC takes precedence over PPLA)|
|P|PPLA2|seat of a second-order administrative division||
|P|PPLA3|seat of a third-order administrative division||
|P|PPLA4|seat of a fourth-order administrative division||
|P|PPLC|capital of a political entity||
}
\keyword{datasets}
