% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lubricity.R
\name{geonamify}
\alias{geonamify}
\title{Calculate the corresponding \code{\link{geoname}} rows for the asciiname query.}
\usage{
geonamify(query, where = NULL, n = 1, ...)
}
\arguments{
\item{query}{The query.}

\item{where}{The named vector of values analogous to the SQL "WHERE" clause.}

\item{n}{The number of allowable fuzzy search results before returning the top result,
otherwise return nothing if exceeded.}

\item{...}{The parameters for \code{\link{agrep}}.}
}
\value{
The rows or \code{data.frame} with 0 rows.
}
\description{
Calculate the corresponding \code{\link{geoname}} rows for the asciiname query.
}
\seealso{
\code{\link{agrep}}
}
