% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neogeonames.R
\docType{package}
\name{neogeonames}
\alias{neogeonames}
\alias{neogeonames-package}
\title{neogeonames: New GeoNames Data Package and Functions in R}
\description{
neogeonames: New GeoNames Data Package and Functions in R

This package provides a subset of GeoNames Gazetteer data and normalization functions.
}
\author{
\strong{Maintainer}: Daniel Negrón \email{dnegron2@gmu.edu}

}
